+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this package will be documented in this file. The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.0.5] - 2026-07-22
+
+
+### Added
+FBX are treated as foreign assets and should not be dirty when exporting.
+Added a preference to ignore the preflight check when exporting and force a build.
+
 ## [1.0.4] - 2026-02-28